+++ src/effects/dither.rs
@@ -4,6 +4,7 @@ mod diffusion;
 mod halftone;
 mod noise;
 mod palette;
+mod riemersma;
 mod threshold;
 
 pub use cells::SamplingMode;
@@ -18,6 +19,7 @@ pub use halftone::{
 };
 pub use noise::{NoiseAlgorithm, NoiseDither};
 pub use palette::{Palette, PaletteMatchMode, PaletteSize};
+pub use riemersma::RiemersmaDither;
 pub use threshold::{OrderedDither, Threshold, ThresholdMap, ThresholdRotation};
 
 #[cfg(test)]